requirements.txt
streamlit>=1.31.0
anthropic>=0.8.0
pandas>=2.0.0
python-dotenv>=1.0.0